+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.110.1
 
 * `[support/pkgs]` Improved formatting
+* `[support/network]` Local IPs removed from IPv4/IPv6 address lists
 
 ### 12.110.0